requirements.txt
torch
torchsde
torchvision
torchaudio
einops
transformers>=4.25.1
safetensors>=0.4.2
aiohttp
pyyaml
Pillow
scipy
tqdm
psutil

#non essential dependencies:
kornia>=0.7.1
spandrel
soundfile

#kidgeni

websocket-client
fastapi
uvicorn
